// src/data/types.ts
export interface TaxonomyType {
  id: string | number;
  name: string;
  href: Route<string> | string | PathName;
  count?: number;
  thumbnail?: string;
  desc?: string;
  color?: TwMainColor | string;
  taxonomy: "category" | "tag";
  listingType?: "stay" | "experiences" | "car" | "last minutes" | "time based";
}
// ...
export type TwMainColor =
  | "pink"
  | "green"
  | "yellow"
  | "red"
  | "indigo"
  | "blue"
  | "purple"
  | "gray";